Refactor transaction displaying using metadata

+++ app/src/parser.h
@@ -38,8 +38,7 @@ parser_error_t parser_validate(const parser_context_t *ctx);
 parser_error_t parser_getNumItems(const parser_context_t *ctx, uint8_t *num_items);
 
 // retrieves a readable output for each field / page
-parser_error_t parser_getItem(const parser_context_t *ctx,
-                              uint16_t displayIdx,
+parser_error_t parser_getItem(const parser_context_t *ctx, uint8_t displayIdx,
                               char *outKey, uint16_t outKeyLen,
                               char *outVal, uint16_t outValLen,
                               uint8_t pageIdx, uint8_t *pageCount);
